+++ lib/flop.ex
@@ -424,7 +424,7 @@ defmodule Flop do
     q
     |> filter(flop, opts)
     |> order_by(flop, opts)
-    |> paginate(flop)
+    |> paginate(flop, opts)
   end
 
   @doc """
@@ -769,7 +769,7 @@ defmodule Flop do
           after: nil,
           offset: nil
         },
-        _opts
+        opts
       )
       when is_integer(last) do
     reversed_order =
@@ -777,15 +777,37 @@ defmodule Flop do
       |> prepare_order(directions)
       |> reverse_ordering()
 
-    Query.order_by(q, ^reversed_order)
+    case opts[:for] do
+      nil ->
+        Query.order_by(q, ^reversed_order)
+
+      module ->
+        struct = struct(module)
+
+        Enum.reduce(reversed_order, q, fn expr, acc_q ->
+          Flop.Schema.apply_order_by(struct, acc_q, expr)
+        end)
+    end
   end
 
   def order_by(
         q,
         %Flop{order_by: fields, order_directions: directions},
-        _opts
+        opts
       ) do
-    Query.order_by(q, ^prepare_order(fields, directions))
+    case opts[:for] do
+      nil ->
+        Query.order_by(q, ^prepare_order(fields, directions))
+
+      module ->
+        struct = struct(module)
+
+        fields
+        |> prepare_order(directions)
+        |> Enum.reduce(q, fn expr, acc_q ->
+          Flop.Schema.apply_order_by(struct, acc_q, expr)
+        end)
+    end
   end
 
   @spec prepare_order([atom], [order_direction()]) :: [
@@ -821,8 +843,10 @@ defmodule Flop do
 
   Used by `Flop.query/2`.
   """
-  @spec paginate(Queryable.t(), Flop.t()) :: Queryable.t()
-  def paginate(q, %Flop{limit: limit, offset: offset})
+  @spec paginate(Queryable.t(), Flop.t(), keyword) :: Queryable.t()
+  def paginate(q, flop, opts \\ [])
+
+  def paginate(q, %Flop{limit: limit, offset: offset}, _)
       when (is_integer(limit) and limit >= 1) or
              (is_integer(offset) and offset >= 0) do
     q
@@ -830,7 +854,7 @@ defmodule Flop do
     |> offset(offset)
   end
 
-  def paginate(q, %Flop{page: page, page_size: page_size})
+  def paginate(q, %Flop{page: page, page_size: page_size}, _)
       when is_integer(page) and is_integer(page_size) and
              page >= 1 and page_size >= 1 do
     q
@@ -838,13 +862,17 @@ defmodule Flop do
     |> offset((page - 1) * page_size)
   end
 
-  def paginate(q, %Flop{
-        first: first,
-        after: nil,
-        before: nil,
-        last: nil,
-        limit: nil
-      })
+  def paginate(
+        q,
+        %Flop{
+          first: first,
+          after: nil,
+          before: nil,
+          last: nil,
+          limit: nil
+        },
+        _
+      )
       when is_integer(first),
       do: limit(q, first + 1)
 
@@ -858,13 +886,14 @@ defmodule Flop do
           before: nil,
           last: nil,
           limit: nil
-        }
+        },
+        opts
       )
       when is_integer(first) do
     orderings = prepare_order(order_by, order_directions)
 
     q
-    |> apply_cursor(after_, orderings)
+    |> apply_cursor(after_, orderings, opts)
     |> limit(first + 1)
   end
 
@@ -878,7 +907,8 @@ defmodule Flop do
           first: nil,
           after: nil,
           limit: nil
-        }
+        },
+        opts
       )
       when is_integer(last) do
     prepared_order_reversed =
@@ -887,11 +917,11 @@ defmodule Flop do
       |> reverse_ordering()
 
     q
-    |> apply_cursor(before, prepared_order_reversed)
+    |> apply_cursor(before, prepared_order_reversed, opts)
     |> limit(last + 1)
   end
 
-  def paginate(q, _), do: q
+  def paginate(q, _, _), do: q
 
   ## Offset/limit pagination
 
@@ -905,45 +935,72 @@ defmodule Flop do
 
   ## Cursor pagination helpers
 
-  @spec apply_cursor(Queryable.t(), map() | nil, [order_direction()]) ::
-          Queryable.t()
-  defp apply_cursor(q, nil, _), do: q
+  @spec apply_cursor(
+          Queryable.t(),
+          map() | nil,
+          [order_direction()],
+          keyword
+        ) :: Queryable.t()
+  defp apply_cursor(q, nil, _, _), do: q
 
-  defp apply_cursor(q, cursor, ordering) do
+  defp apply_cursor(q, cursor, ordering, opts) do
     cursor = Cursor.decode!(cursor)
-    where_dynamic = cursor_dynamic(ordering, cursor)
+
+    where_dynamic =
+      case opts[:for] do
+        nil ->
+          cursor_dynamic(ordering, cursor)
+
+        module ->
+          module
+          |> struct()
+          |> Flop.Schema.cursor_dynamic(ordering, cursor)
+      end
+
     Query.where(q, ^where_dynamic)
   end
 
-  defp cursor_dynamic([], _), do: nil
+  defp cursor_dynamic([], _), do: true
 
   defp cursor_dynamic([{direction, field}], cursor) do
-    case direction do
-      dir when dir in [:asc, :asc_nulls_first, :asc_nulls_last] ->
-        Query.dynamic([r], field(r, ^field) > ^cursor[field])
+    field_cursor = cursor[field]
 
-      dir when dir in [:desc, :desc_nulls_first, :desc_nulls_last] ->
-        Query.dynamic([r], field(r, ^field) < ^cursor[field])
+    if is_nil(field_cursor) do
+      true
+    else
+      case direction do
+        dir when dir in [:asc, :asc_nulls_first, :asc_nulls_last] ->
+          Query.dynamic([r], field(r, ^field) > ^field_cursor)
+
+        dir when dir in [:desc, :desc_nulls_first, :desc_nulls_last] ->
+          Query.dynamic([r], field(r, ^field) < ^field_cursor)
+      end
     end
   end
 
   defp cursor_dynamic([{direction, field} | [{_, _} | _] = tail], cursor) do
     field_cursor = cursor[field]
 
-    case direction do
-      dir when dir in [:asc, :asc_nulls_first, :asc_nulls_last] ->
-        Query.dynamic(
-          [r],
-          field(r, ^field) >= ^field_cursor and
-            (field(r, ^field) > ^field_cursor or ^cursor_dynamic(tail, cursor))
-        )
+    if is_nil(field_cursor) do
+      cursor_dynamic(tail, cursor)
+    else
+      case direction do
+        dir when dir in [:asc, :asc_nulls_first, :asc_nulls_last] ->
+          Query.dynamic(
+            [r],
+            field(r, ^field) >= ^field_cursor and
+              (field(r, ^field) > ^field_cursor or
+                 ^cursor_dynamic(tail, cursor))
+          )
 
-      dir when dir in [:desc, :desc_nulls_first, :desc_nulls_last] ->
-        Query.dynamic(
-          [r],
-          field(r, ^field) <= ^field_cursor and
-            (field(r, ^field) < ^field_cursor or ^cursor_dynamic(tail, cursor))
-        )
+        dir when dir in [:desc, :desc_nulls_first, :desc_nulls_last] ->
+          Query.dynamic(
+            [r],
+            field(r, ^field) <= ^field_cursor and
+              (field(r, ^field) < ^field_cursor or
+                 ^cursor_dynamic(tail, cursor))
+          )
+      end
     end
   end
 
@@ -1037,7 +1094,7 @@ defmodule Flop do
       iex> msg
       "has an invalid entry"
       iex> enum
-      [:name, :age]
+      [:name, :age, :owner_name, :owner_age]
 
   Note that currently, trying to use an existing field that is not allowed as
   seen above will result in the error message `has an invalid entry`, while